Рефакторинг + Добавление выбора компаний
