+++ config.ini
@@ -3,5 +3,5 @@ token = 1245516512:AAFAPDaDe2DwPgqTZxy4eanjSpLd5VigsUg
 webhook_url = https://health-meters-bot.herokuapp.com
 
 [debug]
-webhook_url = https://15574b620f45.ngrok.io
+webhook_url = https://29492456f6ca.ngrok.io
 server_url = http://127.0.0.1:8000/telegram_bot
